+++ Zadatak/src/main/java/com/zadatak/domain/City.java
@@ -19,7 +19,7 @@ public class City {
 	@Id
 	@GeneratedValue
 	@Column(name = "ID")
-	private int id;
+	private Long id;
 	
 	@Size(max = 50, message = "* Max 50 letters.")
 	@NotEmpty(message = "* Cannot be empty")
@@ -34,18 +34,18 @@ public class City {
 
 	public City() {}
 
-	public City(int id, String name, String zipCode, Country country) {
+	public City(Long id, String name, String zipCode, Country country) {
 		this.id = id;
 		this.name = name;
 		this.zipCode = zipCode;
 		this.country = country;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
