+++ Zadatak/src/main/java/com/zadatak/domain/Sex.java
@@ -14,20 +14,20 @@ public class Sex {
 	@Id
 	@GeneratedValue
 	@Column(name = "ID")
-	private int id;
+	private Long id;
 	
 	private String name;
 
-	public Sex(int id, String name) {
+	public Sex(Long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
